Dependency Analysis engine for project
It is charged with finding all types and/or packages that are dependent
on a specified type or package.  It must do this for source code file
sets that may reside on one or more directories in local or remote
machines.  This will require you to develop both server and client
programs where the server may reside on a remote machine.